Admin scherm bijna klaar; enkel nog voor update nakijken.

+++ HamzaIndividueelProject/Inventory.cs
@@ -11,7 +11,7 @@ namespace HamzaIndividueelProject
         public int Id { get; set; }
         public string Name { get; set; }
         public string ModelName { get; set; }
-        public int RetailPrice { get; set; }
+        public double RetailPrice { get; set; }
         public int Quantity { get; set; }
     }
 }
